# Conflicts:
#	libs/vm/include/vm/module.hpp
#	libs/vm/include/vm/module/class_constructor_export.hpp
#	libs/vm/include/vm/module/class_member_export.hpp
#	libs/vm/include/vm/module/function_export.hpp
#	libs/vm/include/vm/module/stack_loader.hpp
#	libs/vm/include/vm/vm.hpp
#	libs/vm/src/analyser.cpp
#	libs/vm/src/string.cpp

+++ libs/core/include/core/random/lfg.hpp
@@ -1,7 +1,7 @@
 #pragma once
 //------------------------------------------------------------------------------
 //
-//   Copyright 2018 Fetch.AI Limited
+//   Copyright 2018-2019 Fetch.AI Limited
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -96,7 +96,8 @@ private:
   LinearCongruentialGenerator lcg_;
 
   random_type             buffer_[Q];
-  static constexpr double inv_double_max_ = 1. / std::numeric_limits<random_type>::max();
+  static constexpr double inv_double_max_ =
+      1. / static_cast<double>(std::numeric_limits<random_type>::max());
 };
 }  // namespace random
 }  // namespace fetch
